actually submit events
