+++ packages/api/src/controllers/webhooksController.ts
@@ -4,6 +4,7 @@ import {
   generateDigest,
   verifyTimestampedSignature,
 } from "backend-lib/src/crypto";
+import { submitSendgridEvents } from "backend-lib/src/destinations/sendgrid";
 import logger from "backend-lib/src/logger";
 import prisma from "backend-lib/src/prisma";
 import { SendgridEvent } from "backend-lib/src/types";
@@ -88,6 +89,10 @@ export default async function webhookController(fastify: FastifyInstance) {
         });
       }
 
+      await submitSendgridEvents({
+        workspaceId,
+        events: request.body,
+      });
       return reply.status(200).send();
     }
   );
